app/requirements.txt
Django==3.1.2
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
requests==2.24.0
beautifulsoup4==4.9.3
